requirements.txt
alembic==1.13.2
aniso8601==9.0.1
astroid==3.2.4
attrs==24.2.0
blinker==1.8.2
click==8.1.7
dill==0.3.8
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
importlib_resources==6.4.2
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycodestyle==2.12.1
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
rpds-py==0.20.0
SQLAlchemy==2.0.32
tomli==2.0.1
tomlkit==0.13.2
typing_extensions==4.12.2
Werkzeug==3.0.3
